fix: 공고관리와 학생의 지원 기능에 맞게 컨트롤러 분리
